nss-reservation-service/src/main/java/cz/cvut/reservation/dao/UserDao.java
package cz.cvut.reservation.dao;

import cz.cvut.reservation.model.User;
import org.springframework.stereotype.Repository;

import javax.persistence.NoResultException;

@Repository
public class UserDao extends BaseDao<User> {

    public UserDao() {
        super(User.class);
    }

    public User findByUsername(String username) {
        try {
            return em.createNamedQuery("User.findByUsername", User.class).setParameter("username", username)
                     .getSingleResult();
        } catch (NoResultException e) {
            throw new NoResultException("User with username " + username + " does not exist.");
        }
    }
}
